Reorganized /app to have sub-components nested within parent components (i.e. about-content is nested within about).
Implemented a user agent-based browser tester and platform tester to find when safari mobile is being used. In this case, the footer is heightened 64px so that the mobile Safari browser menu will not appear when the user taps in the lower 44px of the screen while in expanded view mode. This css is tweak only exists in Safari mobile.